+++ src/components/VideoProvider/index.tsx
@@ -6,6 +6,12 @@
  *
  * Aug 6, 2020 WQP
  * - updated to latest twilio-video starter app
+ *
+ * Aug 27, 2020 WQP
+ * - Add onError to useLocalTracks
+ *
+ * Aug 30, 2020 WQP
+ * - Add AttachVisibilityHandler, per starter app
  */
 import React, { createContext, ReactNode } from 'react';
 import {
@@ -20,8 +26,7 @@ import {
 import { Callback, ErrorCallback } from '../../types';
 import { SelectedParticipantProvider } from './useSelectedParticipant/useSelectedParticipant';
 
-// TODO, this is a good idea and we should probably do it https://github.com/twilio/twilio-video-app-react/blob/master/src/components/VideoProvider/AttachVisibilityHandler/AttachVisibilityHandler.tsx
-// import AttachVisibilityHandler from './AttachVisibilityHandler/AttachVisibilityHandler';
+import AttachVisibilityHandler from './AttachVisibilityHandler/AttachVisibilityHandler';
 import useHandleRoomDisconnectionErrors from './useHandleRoomDisconnectionErrors/useHandleRoomDisconnectionErrors';
 import useHandleOnDisconnect from './useHandleOnDisconnect/useHandleOnDisconnect';
 import useHandleTrackPublicationFailed from './useHandleTrackPublicationFailed/useHandleTrackPublicationFailed';
@@ -95,6 +100,11 @@ export function VideoProvider({ options, children, onError = () => {}, onDisconn
     >
       <RoomStateProvider>
         <SelectedParticipantProvider room={room}>{children}</SelectedParticipantProvider>
+        {/*
+          The AttachVisibilityHandler component is using the useLocalVideoToggle hook
+          which must be used within the VideoContext Provider.
+        */}
+        <AttachVisibilityHandler />
       </RoomStateProvider>
     </VideoContext.Provider>
   );